+++ lobbuble/app/src/main/AndroidManifest.xml
@@ -12,7 +12,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
 
-        <activity android:name=".MainActivity">
+        <activity android:name=".activity.MainActivity">
         </activity>
         <activity android:name=".activity.SplashActivity">
             <intent-filter>
