--- lobbuble/app/src/main/java/net/jspiner/lobbuble/MainActivity.java
@@ -1,67 +1,0 @@
-package net.jspiner.lobbuble;
-
-import android.content.Intent;
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-import android.util.Log;
-
-import com.facebook.CallbackManager;
-import com.facebook.FacebookCallback;
-import com.facebook.FacebookException;
-import com.facebook.login.LoginResult;
-import com.facebook.login.widget.LoginButton;
-
-public class MainActivity extends AppCompatActivity {
-
-    private static String TAG = MainActivity.class.getSimpleName();
-
-    CallbackManager callbackManager;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-
-        init();
-    }
-
-    void init(){
-
-        final LoginButton loginButton = (LoginButton)findViewById(R.id.login_button);
-
-        callbackManager = CallbackManager.Factory.create();
-
-        loginButton.setReadPermissions("email");
-        loginButton.registerCallback(callbackManager, new FacebookCallback<LoginResult>() {
-            @Override
-            public void onSuccess(LoginResult loginResult) {
-                Log.d(TAG, "onSuccess");
-
-                String loginToken = loginResult.getAccessToken().getToken();
-                Log.d(TAG, "token : " + loginToken);
-            }
-
-            @Override
-            public void onCancel() {
-                Log.d(TAG, "onCancel");
-
-
-            }
-
-            @Override
-            public void onError(FacebookException error) {
-                Log.d(TAG, "onError");
-                Log.e(TAG, "error msg : " + error.getMessage());
-
-            }
-        });
-
-
-    }
-
-    @Override
-    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
-        super.onActivityResult(requestCode, resultCode, data);
-        callbackManager.onActivityResult(requestCode, resultCode, data);
-    }
-}
