+++ lobbuble/app/src/main/java/net/jspiner/lobbuble/activity/MainActivity.java
@@ -1,0 +1,33 @@
+package net.jspiner.lobbuble.activity;
+
+import android.content.Intent;
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+import android.util.Log;
+
+import com.facebook.CallbackManager;
+import com.facebook.FacebookCallback;
+import com.facebook.FacebookException;
+import com.facebook.login.LoginResult;
+import com.facebook.login.widget.LoginButton;
+
+import net.jspiner.lobbuble.R;
+import net.jspiner.lobbuble.activity.base.BaseActivity;
+
+public class MainActivity extends BaseActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+
+        init();
+    }
+
+    void init(){
+
+
+
+    }
+
+}
