+++ lobbuble/app/src/main/java/net/jspiner/lobbuble/contract/base/BaseView.java
@@ -8,4 +8,5 @@ public interface BaseView<T> {
 
     void setPresenter(T presenter);
 
+    void showToast(String text);
 }
