+++ lobbuble/app/src/main/res/layout/activity_main.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="net.jspiner.lobbuble.MainActivity">
+    tools:context="net.jspiner.lobbuble.activity.MainActivity">
 
     <TextView
         android:layout_width="wrap_content"
